requirements.txt
paho-mqtt==2.x
moviepy
requests